apps/user_service/src/requirement.txt
astroid==2.2.4
autopep8==1.4.3
grpcio==1.19.0
grpcio-tools==1.19.0
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.7.0
pycodestyle==2.5.0
pylint==2.3.1
pymongo==3.7.2
six==1.12.0
typed-ast==1.3.1
uuid==1.30
wrapt==1.11.1
